# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Synapta API",
    "description": "Event-oriented AI generation backend. Manages events, programs, pieces, AI content generation, files, ingredients, presets, skills, and agency orchestration.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:{port}",
      "variables": {
        "port": { "default": "9099" }
      }
    }
  ],
...
  "paths": {
...
    "/api/v1/agency/sync-asset": {
      "post": {
        "tags": ["Agency Core"],
        "summary": "Sync an external asset to a model",
        "operationId": "syncAsset",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SyncAssetRequest" } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } } }
      }
    },
...
    "/api/v1/agency/logs/generation/{id}": {
      "get": {
        "tags": ["Agency Core"],
        "summary": "Get generation log",
        "operationId": "getGenerationLog",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } } }
      }
    },
...
    "/api/v1/agency/logs/server-communications/{id}": {
      "get": {
        "tags": ["Agency Core"],
        "summary": "Get server communication log",
        "operationId": "getServerComm",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiResponse" } } } } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from POST /auth/login. Carries sub (user id), tenant_id, role_level, username."
      }
    },
    "schemas": {
...
      "SyncAssetRequest": {
        "type": "object",
        "properties": {
          "model_id": { "type": "string", "format": "uuid" },
          "file_id": { "type": "string", "format": "uuid" },
          "metadata": { "type": "object" }
        }
      },
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "message": { "type": "string" },
          "data": { "nullable": true }
        }
      },
...
    },
...
  }
}